% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_chris_fact.R
\name{generate_chris_fact}
\alias{generate_chris_fact}
\title{Chris Halsch fact generator}
\usage{
generate_chris_fact(spicyness)
}
\arguments{
\item{spicyness}{An integer between 1 and 5 describing how spicy you want the fact to be}
}
\value{
A character vector of 1 with a fact about Chris in quotes
}
\description{
Generates a random fact about Chris Halsch with a specified degree of spicyness
}
\examples{
generate_chris_fact(spicyness = 3)

}
