% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_feeds.R
\name{get_mean_feeds}
\alias{get_mean_feeds}
\title{Get mean feeds for a brood}
\usage{
get_mean_feeds(brood_id_I_want)
}
\arguments{
\item{brood_id_I_want}{A character vector of length one with the brood id}
}
\value{
Logical vector of size 1 with the mean number of observed feeds
}
\description{
Gets the mean number of observed feeds for a specific brood id within the "meerkats" dataset.
}
\examples{
get_mean_feeds("bfb0802")

}
