% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_avengers.R
\name{match_avengers}
\alias{match_avengers}
\title{Match avengers}
\usage{
match_avengers(avg_1, avg_2)
}
\arguments{
\item{avg_1}{Character vector of length one with the name of an avenger to be matched}

\item{avg_2}{Character vector of length one with the name of an avenger to be matched}
}
\value{
Character vector of length one with the name of the winner
}
\description{
Predicts who would win in a physical fight between two avengers based on their power
}
\examples{
match_avengers("vision", "scarlett witch")

}
