% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chris_fact.R, R/time_til_val_group.R
\name{time_till_val_group}
\alias{time_till_val_group}
\title{Time until Valentina makes a group chat}
\usage{
time_till_val_group(n_people)

time_till_val_group(n_people)
}
\arguments{
\item{n_people}{Positive integer setting the number of participants in the social event.}
}
\value{
A logical vector of size 1 describing the time it takes in seconds
}
\description{
Predicts the time it will take for Valentina to make a group chat for a social event depending on the number of people taking part in it.

Predicts the time it will take for Valentina to make a group chat for a social event depending on the number of people taking part in it.
}
\examples{
time_till_val_group(12)

time_till_val_group(12)

}
